[BAEL-4381] Intro to ArchUnit
Package typo
