+++ libraries-testing/src/main/java/com/baeldung/archunit/smurfs/persistence/SmurfsRepository.java
@@ -1,7 +1,7 @@
 /**
  * 
  */
-package com.baldung.archunit.smurfs.persistence;
+package com.baeldung.archunit.smurfs.persistence;
 
 import java.util.Collections;
 import java.util.List;
@@ -9,7 +9,7 @@ import java.util.Map;
 import java.util.Optional;
 import java.util.TreeMap;
 
-import com.baldung.archunit.smurfs.persistence.domain.Smurf;
+import com.baeldung.archunit.smurfs.persistence.domain.Smurf;
 
 import static java.util.stream.Collectors.toList;
