+++ libraries-testing/src/main/java/com/baeldung/archunit/smurfs/presentation/SmurfsController.java
@@ -1,4 +1,4 @@
-package com.baldung.archunit.smurfs.presentation;
+package com.baeldung.archunit.smurfs.presentation;
 
 import java.util.List;
 
@@ -7,8 +7,8 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.baldung.archunit.smurfs.service.SmurfsService;
-import com.baldung.archunit.smurfs.service.dto.SmurfDTO;
+import com.baeldung.archunit.smurfs.service.SmurfsService;
+import com.baeldung.archunit.smurfs.service.dto.SmurfDTO;
 
 @RequestMapping(value = "/smurfs", produces = MediaType.APPLICATION_JSON_UTF8_VALUE)
 @RestController
